Correct some of kernel
